% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tindR.R
\name{tindR}
\alias{tindR}
\title{Swipe on the columns of your data frame}
\usage{
tindR(df, by = NULL)
}
\arguments{
\item{df}{the data frame to select columns from}

\item{by}{optional column to group by}
}
\description{
Select columns from a data frame in an interactive manner
}
